+++ pyp5js/cli.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python3
+import warnings
 from pathlib import Path
 
 from cprint import cprint
 import click
 
 from pyp5js import commands
-from pyp5js.config import SKETCHBOOK_DIR
+from pyp5js.config import SKETCHBOOK_DIR, AVAILABLE_INTERPRETERS, PYODIDE_INTERPRETER
 
 
 @click.group()
@@ -18,34 +19,26 @@ def command_line_entrypoint():
 
 @command_line_entrypoint.command('new')
 @click.argument('sketch_name')
-@click.option('--monitor', '-m', is_flag=True)
-def configure_new_sketch(sketch_name, monitor):
+@click.option('--monitor', '-m', is_flag=True, help='Starts the monitor command too')
+@click.option('--interpreter', '-i', type=click.Choice(AVAILABLE_INTERPRETERS), default=PYODIDE_INTERPRETER, help='Which python tool to use to run the sketch. (defaults to pyodide)')
+def configure_new_sketch(sketch_name, monitor, interpreter):
     """
-    Create dir and configure boilerplate
-
-    Params:
-    - sketch_name: name of the sketch (will create a {sketch_name}.py)
-
-    Opitionals:
-    - monitor: start the monitor command as well
-
-    Example:
-    $ pyp5js new my_sketch
+    Create dir and configure boilerplate - Example:\n
+    $ pyp5js new my_sketch -i pyodide
     """
-    files = commands.new_sketch(sketch_name)
+    files = commands.new_sketch(sketch_name, interpreter)
 
     cprint.ok(f"Your sketch was created!")
 
     if not monitor:
         cprint.ok(f"Please, open and edit the file {files.sketch_py} to draw. When you're ready to see your results, just run:")
-        cmd = f"\t pyp5js transcrypt {sketch_name}"
+        cmd = f"\t pyp5js compile {sketch_name}"
         cprint.ok(cmd)
         cprint.ok(f"And open file://{files.index_html.absolute()} on your browser to see yor results!")
     else:
-        cprint.ok(f"Please, open and edit the file {sketch_py} to draw.")
+        cprint.ok(f"Please, open and edit the file {files.sketch_py} to draw.")
         cprint.ok(f"And open file://{files.index_html.absolute()} on your browser to see yor results!")
         commands.monitor_sketch(sketch_name)
-
 
 
 @command_line_entrypoint.command("transcrypt")
@@ -55,12 +48,20 @@ def transcrypt_sketch(sketch_name):
     Command to generate the P5.js code for a python sketch
 
     Params:
+
     - sketch_name: name of the sketch
 
     Example:
     $ pyp5js transcrypt my_sketch
     """
-    files = commands.transcrypt_sketch(sketch_name)
+    msg = f"transcript command is deprecated. Instead, please run: \n\n\tpyp5js compile {sketch_name}\n"
+    warnings.warn(msg, UserWarning)
+
+
+@command_line_entrypoint.command("compile")
+@click.argument("sketch_name")
+def compile_sketch(sketch_name):
+    files = commands.compile_sketch(sketch_name)
     cprint.ok(f"Your sketch is ready and available at file://{files.index_html.absolute()}")
 
 
@@ -72,9 +73,11 @@ def monitor_sketch(sketch_name):
     it'll automatically generate the JS files as in pyp5js transcrypt command
 
     Params:
+
     - sketch_name: name of the sketch
 
     Example:
+
     $ pyp5js monitor my_sketch
     """
     commands.monitor_sketch(sketch_name)